BREAKING CHANGE: CommandHandler, new bot structure
