+++ env.d.ts
@@ -2,7 +2,7 @@ declare global {
   namespace NodeJS {
     interface ProcessEnv {
       BOT_TOKEN: string;
-      IS_DEV: boolean;
+      IS_DEV: string;
       ADMIN_ID: number;
       NODE_ENV: 'development' | 'production';
       PORT?: string;
