+++ package.json
@@ -1,18 +1,20 @@
 {
   "name": "schedulevgtubot",
-  "version": "4.0.0",
+  "version": "5.0.0",
   "description": "",
   "main": "index.js",
   "type": "module",
   "scripts": {
     "start": "node dist/index.js",
-    "build": "tsc --build"
+    "build": "tsc --build",
+    "dev": "ts-node-esm src/index.ts"
   },
   "author": "Slincnik",
   "license": "ISC",
   "dependencies": {
     "cron": "^3.1.0",
     "dotenv": "^16.3.1",
+    "glob": "^10.3.10",
     "luxon": "^3.4.3",
     "sqlite": "^5.0.1",
     "sqlite3": "^5.1.6",
