+++ src/hearCommands/scheduleNow.ts
@@ -1,0 +1,24 @@
+import { CommandHandler } from '../structures/command.js';
+import {
+  loadScheduleAndReturnAll,
+  parityWeek,
+  returnCouplesMessage,
+  returnScheduleFromDayOfWeek,
+} from '../utils/utils.js';
+
+export default new CommandHandler({
+  name: '👁 Schedule',
+  async execute(ctx) {
+    const { scheduleJson, weekNumber, dayOfWeek, parity } = loadScheduleAndReturnAll();
+
+    const findedSchedule = returnScheduleFromDayOfWeek(scheduleJson, dayOfWeek, parity, weekNumber);
+
+    if (!findedSchedule?.length) {
+      return ctx.reply('Сегодня занятий нету');
+    }
+
+    ctx.reply(`🔷🔷 ${dayOfWeek} (${parityWeek[parity]}) 🔷🔷
+  ${findedSchedule.map((value) => returnCouplesMessage(value)).join('\n\n')}
+  `);
+  },
+});
